+++ lib/auto_start_flutter_method_channel.dart
@@ -11,7 +11,8 @@ class MethodChannelAutoStartFlutter extends AutoStartFlutterPlatform {
 
   @override
   Future<String?> getPlatformVersion() async {
-    final version = await methodChannel.invokeMethod<String>('getPlatformVersion');
+    final version =
+        await methodChannel.invokeMethod<String>('getPlatformVersion');
     return version;
   }
 }